SLIDE refactor to prepare for more advanced movement
instead of hardcoding slide as being an initial 99 moves, instead code slide like this: After we move, if we succeeded and we're currently slide, try to move one more time. Add on top of that some logic to prevent computational explosions (repeat 99 times max, don't insert a unitid twice in still_moving) and we have all the same behaviour as before (except now X on Y is slide doesn't go arbitrarily far, which is cute).